fix: Update model field aliases to match stable API response format
The stable API returns different field names than the old v3/v4 API:
- Quote: changesPercentage -> changePercentage (CRITICAL: caused 404s)
- Quote/Profile: marketCap is now a float, not int
- Profile: mktCap -> marketCap, volAvg -> averageVolume,
  lastDiv -> lastDividend, exchangeShortName -> exchange
- Fundamentals: fillingDate -> filingDate, calendarYear -> fiscalYear
- AftermarketQuote: completely new bid/ask schema

+++ fmp_data_client/server/models.py
@@ -31,7 +31,7 @@ class QuoteResponse(BaseModel):
     day_high: float
     day_low: float
     previous_close: float
-    market_cap: Optional[int] = None
+    market_cap: Optional[float] = None
     timestamp: Optional[int] = None
 
 
@@ -47,7 +47,7 @@ class ProfileResponse(BaseModel):
     employees: Optional[int] = None
     website: Optional[str] = None
     country: Optional[str] = None
-    market_cap: Optional[int] = None
+    market_cap: Optional[float] = None
     exchange: Optional[str] = None
